refactor(discovery): remove unused fingerprinting code and implement optimal TTL calculation
